requirements.txt
azure-functions
requests
line-bot-sdk
yfinance
pandas
BeautifulSoup4
twstock